requirements.txt
PyQt5==5.15.10
mysql-connector-python==8.2.0
GitPython==3.1.40
psutil==5.9.6
pyyaml==6.0.1
requests==2.31.0